+++ cli/src/change/router.rs
@@ -159,8 +159,11 @@ impl CliCommand for RouterCommand {
             matches,
             "router name",
             None,
-            |input: &str| validate_name(input),
-            |_| "Router name cannot be empty or include spaces. Please try again".to_string(),
+            |input: &str| validate_name(input) && !manifest_data.app_name.contains(input),
+            |_| {
+                "Router name cannot be a substring of the application name, empty or include numbers or spaces. Please try again"
+                    .to_string()
+            },
         )?;
 
         let mut removal_templates = vec![];