+++ cli/src/core/package_json/package_json_constants.rs
@@ -1,6 +1,9 @@
 use std::collections::HashSet;
 
-use crate::constants::{Database, Formatter, Linter, Runtime, TestFramework};
+use crate::{
+    constants::{Database, Formatter, Linter, Runtime, TestFramework},
+    core::database::is_in_memory_database,
+};
 
 // Shared package.json devDpendencies constants
 // @biomejs/biome
@@ -18,7 +21,7 @@ pub(crate) const PRETTIER_VERSION: &str = "^3.5.3";
 // ts-jest
 pub(crate) const TS_JEST_VERSION: &str = "^29.2.6";
 // typescript-eslint
-pub(crate) const TYPESCRIPT_ESLINT_VERSION: &str = "^8.30.1";
+pub(crate) const TYPESCRIPT_ESLINT_VERSION: &str = "^8.33.0";
 // vitest
 pub(crate) const VITEST_VERSION: &str = "^3.0.8";
 
@@ -34,7 +37,7 @@ pub(crate) const NODE_GYP_VERSION: &str = "^11.2.0";
 // sort-package-json
 pub(crate) const SORT_PACKAGE_JSON_VERSION: &str = "^3.0.0";
 // tsx
-pub(crate) const TSX_VERSION: &str = "^4.19.3";
+pub(crate) const TSX_VERSION: &str = "^4.19.4";
 // typescript
 pub(crate) const TYPESCRIPT_VERSION: &str = "^5.8.2";
 
@@ -128,6 +131,7 @@ pub(crate) fn application_migrate_script<'a>(
 
     let mut databases = databases
         .iter()
+        .filter(|db| !is_in_memory_database(db))
         .map(|db| db.to_string())
         .collect::<Vec<String>>();
     if let Some(db_init) = db_init {
@@ -135,7 +139,8 @@ pub(crate) fn application_migrate_script<'a>(
     }
 
     format!(
-        "docker compose up -d {} && {}{} run migrate:{}",
+        "docker compose up{}{} && {}{} run migrate:{}",
+        if databases.len() > 0 { " -d " } else { "" },
         databases.join(" -d "),
         sleep,
         package_manager,
@@ -205,51 +210,51 @@ pub(crate) const APP_CORE_VERSION: &str = "workspace:*";
 // @forklaunch/blueprint-monitoring
 pub(crate) const APP_MONITORING_VERSION: &str = "workspace:*";
 // @forklaunch/common
-pub(crate) const COMMON_VERSION: &str = "^0.2.11";
+pub(crate) const COMMON_VERSION: &str = "^0.3.2";
 // @forklaunch/core
-pub(crate) const CORE_VERSION: &str = "^0.7.4";
+pub(crate) const CORE_VERSION: &str = "^0.8.2";
 // @forklaunch/express
-pub(crate) const EXPRESS_VERSION: &str = "^0.4.11";
+pub(crate) const EXPRESS_VERSION: &str = "^0.5.2";
 // @forklaunch/hyper-express
-pub(crate) const HYPER_EXPRESS_VERSION: &str = "^0.4.11";
+pub(crate) const HYPER_EXPRESS_VERSION: &str = "^0.5.2";
 // @forklaunch/implementation-billing-base
-pub(crate) const BILLING_BASE_VERSION: &str = "^0.1.13";
+pub(crate) const BILLING_BASE_VERSION: &str = "^0.1.15";
 // @forklaunch/interfaces-billing
-pub(crate) const BILLING_INTERFACES_VERSION: &str = "^0.1.13";
+pub(crate) const BILLING_INTERFACES_VERSION: &str = "^0.1.15";
 // @forklaunch/implementation-iam-base
-pub(crate) const IAM_BASE_VERSION: &str = "^0.1.13";
+pub(crate) const IAM_BASE_VERSION: &str = "^0.1.15";
 // @forklaunch/interfaces-iam
-pub(crate) const IAM_INTERFACES_VERSION: &str = "^0.1.13";
+pub(crate) const IAM_INTERFACES_VERSION: &str = "^0.1.15";
 // @forklaunch/implementation-worker-bullmq
-pub(crate) const WORKER_BULLMQ_VERSION: &str = "^0.1.5";
+pub(crate) const WORKER_BULLMQ_VERSION: &str = "^0.1.7";
 // @forklaunch/implementation-worker-redis
-pub(crate) const WORKER_REDIS_VERSION: &str = "^0.1.5";
+pub(crate) const WORKER_REDIS_VERSION: &str = "^0.1.7";
 // @forklaunch/implementation-worker-database
-pub(crate) const WORKER_DATABASE_VERSION: &str = "^0.1.5";
+pub(crate) const WORKER_DATABASE_VERSION: &str = "^0.1.7";
 // @forklaunch/implementation-worker-kafka
-pub(crate) const WORKER_KAFKA_VERSION: &str = "^0.1.5";
+pub(crate) const WORKER_KAFKA_VERSION: &str = "^0.1.7";
 // @forklaunch/interfaces-worker
-pub(crate) const WORKER_INTERFACES_VERSION: &str = "^0.1.4";
+pub(crate) const WORKER_INTERFACES_VERSION: &str = "^0.1.6";
 // @forklaunch/validator
-pub(crate) const VALIDATOR_VERSION: &str = "^0.5.4";
+pub(crate) const VALIDATOR_VERSION: &str = "^0.6.2";
 // @mikro-orm/core
-pub(crate) const MIKRO_ORM_CORE_VERSION: &str = "^6.4.13";
+pub(crate) const MIKRO_ORM_CORE_VERSION: &str = "^6.4.16";
 // @mikro-orm/migrations
-pub(crate) const MIKRO_ORM_MIGRATIONS_VERSION: &str = "^6.4.13";
+pub(crate) const MIKRO_ORM_MIGRATIONS_VERSION: &str = "^6.4.16";
 // @mikro-orm/postgresql,@mikro-orm/mongodb,@mikro-orm/mysql,@mikro-orm/better-sqlite,@mikro-orm/sqlite,@mikro-orm/mariadb,@mikro-orm/libsql,@mikro-orm/mssql
-pub(crate) const MIKRO_ORM_DATABASE_VERSION: &str = "^6.4.13";
+pub(crate) const MIKRO_ORM_DATABASE_VERSION: &str = "^6.4.16";
 // @mikro-orm/reflection
-pub(crate) const MIKRO_ORM_REFLECTION_VERSION: &str = "^6.4.13";
+pub(crate) const MIKRO_ORM_REFLECTION_VERSION: &str = "^6.4.16";
 // @mikro-orm/seeder
-pub(crate) const MIKRO_ORM_SEEDER_VERSION: &str = "^6.4.13";
+pub(crate) const MIKRO_ORM_SEEDER_VERSION: &str = "^6.4.16";
 // @sinclair/typebox
 pub(crate) const TYPEBOX_VERSION: &str = "^0.34.33";
 // ajv
 pub(crate) const AJV_VERSION: &str = "^8.17.1";
 // bullmq
-pub(crate) const BULLMQ_VERSION: &str = "^5.52.2";
+pub(crate) const BULLMQ_VERSION: &str = "^5.53.1";
 // better-sqlite3
-pub(crate) const BETTER_SQLITE3_VERSION: &str = "^11.9.1";
+pub(crate) const BETTER_SQLITE3_VERSION: &str = "^11.10.0";
 // dotenv
 pub(crate) const DOTENV_VERSION: &str = "^16.5.0";
 // sqlite3
@@ -257,19 +262,19 @@ pub(crate) const SQLITE3_VERSION: &str = "^5.1.7";
 // uuid
 pub(crate) const UUID_VERSION: &str = "^11.1.0";
 // zod
-pub(crate) const ZOD_VERSION: &str = "^3.24.3";
+pub(crate) const ZOD_VERSION: &str = "^3.25.42";
 
 // Project package.json devDependencies constants
 // @mikro-orm/cli
-pub(crate) const MIKRO_ORM_CLI_VERSION: &str = "^6.4.13";
+pub(crate) const MIKRO_ORM_CLI_VERSION: &str = "^6.4.16";
 // @types/express
-pub(crate) const TYPES_EXPRESS_VERSION: &str = "^5.0.1";
+pub(crate) const TYPES_EXPRESS_VERSION: &str = "^5.0.2";
 // @types/express-serve-static-core
 pub(crate) const TYPES_EXPRESS_SERVE_STATIC_CORE_VERSION: &str = "^5.0.6";
 // @types/qs
-pub(crate) const TYPES_QS_VERSION: &str = "^6.9.18";
+pub(crate) const TYPES_QS_VERSION: &str = "^6.14.0";
 // typedoc
-pub(crate) const TYPEDOC_VERSION: &str = "^0.28.3";
+pub(crate) const TYPEDOC_VERSION: &str = "^0.28.5";
 // @types/uuid
 pub(crate) const TYPES_UUID_VERSION: &str = "^10.0.0";