+++ blueprint/interfaces/billing/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @forklaunch/interfaces-billing
 
+## 0.1.15
+
+### Patch Changes
+
+- increase package versions
+
+## 0.1.14
+
+### Patch Changes
+
+- Upgrade dependencies
+
 ## 0.1.13
 
 ### Patch Changes