+++ framework/core/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @forklaunch/core
 
+## 0.8.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @forklaunch/common@0.3.2
+  - @forklaunch/validator@0.6.2
+
+## 0.8.1
+
+### Patch Changes
+
+- Add additional options to framework to instantiate applications and routers
+- Updated dependencies
+  - @forklaunch/validator@0.6.1
+  - @forklaunch/common@0.3.1
+
+## 0.8.0
+
+### Minor Changes
+
+- Added support for content types in request/response and fixed edge cases in validator
+
+### Patch Changes
+
+- Updated dependencies
+  - @forklaunch/common@0.3.0
+  - @forklaunch/validator@0.6.0
+
 ## 0.7.4
 
 ### Patch Changes